tests: Add integration testing
This tests Cloud Hypervisor and QEMU (PVH) against Ubuntu Bionic/Focal
and Clear Linux.

They are built and run when the "integration_tests" feature is enabled
and expect certain files to be available. This is handled by the
"run_integration_tests.sh" script.

Currently QEMU with Focal is disabled as that does not function.

Signed-off-by: Rob Bradford <robert.bradford@intel.com>